+++ src/elvia/elvia-tariff-types.html
@@ -11,7 +11,7 @@
     icon: "elvia_hvite.svg",
     color: "#FFCC66",
     label: function () {
-      return this.name || "ps-elvia-tariff-types";
+      return this.name || "Elvia Tariff Types";
     },
   });
 </script>
