+++ src/receive-price-functions.js
