+++ src/receive-price.html
@@ -10,7 +10,7 @@
     icon: "font-awesome/fa-eur",
     color: "#FFCC66",
     label: function () {
-      return this.name || "ps-receive-price";
+      return this.name || "Price Receiver";
     },
   });
 </script>
