+++ src/strategy-lowest-price.html
@@ -61,7 +61,7 @@
     icon: "font-awesome/fa-eur",
     color: "#FFCC66",
     label: function () {
-      return this.name || "ps-strategy-lowest-price";
+      return this.name || "Lowest Price";
     },
     outputLabels: ["on", "off", "schedule"],
     oneditprepare: function () {
