+++ test/send-config-input.test.js
@@ -1,6 +1,6 @@
 const expect = require("expect");
 const helper = require("node-red-node-test-helper");
-const bestSave = require("../strategy-best-save.js");
+const bestSave = require("../src/strategy-best-save.js");
 const { DateTime } = require("luxon");
 const prices = require("./data/converted-prices.json");
 const result = require("./data/best-save-result.json");
@@ -111,8 +111,6 @@ describe("send config as input", () => {
             break;
           case 2:
             pass++;
-            console.log("msg.payload:");
-            console.log(msg.payload);
             expect(equalPlan(adjustedResult, msg.payload)).toBeTruthy();
             done();
         }
