+++ test/strategy-best-save.test.js
@@ -2,7 +2,7 @@ const cloneDeep = require("lodash.clonedeep");
 const { DateTime } = require("luxon");
 const expect = require("expect");
 const helper = require("node-red-node-test-helper");
-const bestSave = require("../strategy-best-save.js");
+const bestSave = require("../src/strategy-best-save.js");
 const prices = require("./data/converted-prices.json");
 const result = require("./data/best-save-result.json");
 const { testPlan: plan } = require("./test-utils");
